chore(essentials): deps updated, pa11y and publint added. scripts moved and refactored
