+++ package.json
@@ -30,22 +30,22 @@
 	"devDependencies": {
 		"@changesets/changelog-git": "^0.1.14",
 		"@changesets/cli": "^2.26.0",
-		"@commitlint/cli": "^17.3.0",
-		"@commitlint/config-conventional": "^17.3.0",
+		"@commitlint/cli": "^17.4.1",
+		"@commitlint/config-conventional": "^17.4.0",
 		"@sveltejs/vite-plugin-svelte": "^2.0.2",
 		"@types/gtag.js": "^0.0.12",
 		"@types/node": "^18.11.18",
-		"husky": "^8.0.2",
+		"husky": "^8.0.3",
 		"lint-staged": "^13.1.0",
 		"plop": "^3.1.1",
-		"prettier": "^2.8.1",
+		"prettier": "^2.8.2",
 		"prettier-plugin-svelte": "^2.9.0",
-		"svelte": "^3.55.0",
-		"svelte-check": "^3.0.1",
+		"svelte": "^3.55.1",
+		"svelte-check": "^3.0.2",
 		"svelte-preprocess": "^5.0.0",
-		"turbo": "^1.6.3",
+		"turbo": "^1.7.0",
 		"typescript": "^4.9.4",
-		"vite": "^4.0.3",
-		"vitest": "0.26.2"
+		"vite": "^4.0.4",
+		"vitest": "^0.27.1"
 	}
 }
