+++ packages/essentials/package.json
@@ -3,7 +3,7 @@
 	"version": "0.4.0",
 	"private": false,
 	"type": "module",
-	"description": "A collection of _essentials_ and fully configurable Svelte components.",
+	"description": "A collection of essentials, type-safe, accessible and configurable Svelte components.",
 	"author": "dev <dev@sveltin.io>",
 	"license": "MIT",
 	"homepage": "https://components.sveltin.io",
@@ -11,6 +11,8 @@
 		"sveltekit",
 		"svelte",
 		"components",
+		"accessible components",
+		"a11y components",
 		"buttons",
 		"buttongroup",
 		"color viewer",
@@ -19,7 +21,8 @@
 		"external links",
 		"tabs",
 		"togglelist",
-		"collapse"
+		"collapse",
+		"svelte accessibility"
 	],
 	"repository": {
 		"type": "git",
@@ -30,58 +33,66 @@
 		"url": "https://github.com/sveltinio/components-library/issues"
 	},
 	"dependencies": {
-		"svelte": "^3.55.0"
+		"svelte": "^3.55.1"
 	},
 	"devDependencies": {
 		"@indaco/svelte-iconoir": "^3.1.0",
 		"@sveltejs/adapter-auto": "^1.0.0",
-		"@sveltejs/kit": "^1.0.1",
-		"@sveltejs/package": "^1.0.1",
+		"@sveltejs/kit": "^1.0.11",
+		"@sveltejs/package": "^1.0.2",
 		"@sveltejs/vite-plugin-svelte": "^2.0.2",
 		"@testing-library/jest-dom": "^5.16.5",
 		"@testing-library/svelte": "^3.2.2",
 		"@types/postcss-preset-env": "^7.7.0",
-		"@typescript-eslint/eslint-plugin": "^5.47.1",
-		"@typescript-eslint/parser": "^5.47.1",
+		"@typescript-eslint/eslint-plugin": "^5.48.1",
+		"@typescript-eslint/parser": "^5.48.1",
+		"axe-core": "^4.6.2",
 		"c8": "^7.12.0",
 		"cross-env": "^7.0.3",
 		"cssnano": "^5.1.14",
+		"cssnano-preset-advanced": "^5.3.9",
 		"eslint": "^8.31.0",
 		"eslint-config-custom": "workspace:0.0.0",
 		"jsdom": "^20.0.3",
 		"micromatch": "^4.0.5",
-		"postcss": "^8.4.20",
+		"pa11y": "^6.2.3",
+		"pa11y-ci": "^3.0.1",
+		"picocolors": "^1.0.0",
+		"postcss": "^8.4.21",
 		"postcss-cli": "^10.1.0",
 		"postcss-load-config": "^4.0.1",
 		"postcss-preset-env": "^7.8.3",
 		"postcss-utils": "workspace:0.0.0",
+		"publint": "^0.1.7",
 		"rimraf": "^3.0.2",
 		"stylelint": "^14.16.1",
 		"stylelint-config-custom": "workspace:0.0.0",
-		"svelte-check": "^3.0.1",
-		"svelte-preprocess": "^5.0.0",
+		"svelte-check": "^3.0.2",
 		"tslib": "^2.4.1",
 		"typescript": "^4.9.4",
-		"vite": "^4.0.3",
-		"vitest": "^0.26.3"
+		"vite": "^4.0.4",
+		"vitest": "^0.27.1"
 	},
 	"scripts": {
-		"build": "pnpm clean && pnpm sync && pnpm build:css && svelte-package",
-		"build:css": "cross-env NODE_ENV=production node styles-builder.js",
+		"build": "pnpm clean && pnpm sync && pnpm build:css && svelte-package && pnpm postbuild",
+		"build:css": "cross-env NODE_ENV=production node ./scripts/postcss-process.js",
 		"check": "pnpm sync && svelte-check",
 		"check:watch": "pnpm sync && svelte-check --watch",
 		"clean": "rimraf dist",
 		"dev": "pnpm sync && pnpm dev:css && vite dev",
-		"dev:css": "cross-env NODE_ENV=development node styles-builder.js",
+		"dev:css": "cross-env NODE_ENV=development node ./scripts/postcss-process.js",
 		"lint": "prettier --plugin-search-dir=. --check . && eslint .",
 		"lint:styles": "stylelint \"**/styles.postcss\"",
 		"format": "prettier --plugin-search-dir=. --write .",
+		"minify": "cross-env NODE_ENV=production node ./scripts/css-minifier.js",
+		"postbuild": "pnpm exec publint ./dist && pnpm minify",
 		"sync": "svelte-kit sync",
 		"test": "pnpm sync && vitest run",
 		"test:coverage": "pnpm sync && vitest run --coverage",
-		"test:watch": "pnpm sync && vitest watch"
+		"test:watch": "pnpm sync && vitest watch",
+		"test:a11y": "rimraf a11y-results.json && pa11y-ci"
 	},
-	"types": "index.d.ts",
+	"types": "types.d.ts",
 	"svelte": "index.js",
 	"sideEffects": false,
 	"publishConfig": {
@@ -89,8 +100,8 @@
 	},
 	"exports": {
 		".": {
-			"import": "./index.js",
-			"types": "./types.d.ts"
+			"types": "./types.d.ts",
+			"import": "./index.js"
 		},
 		"./dist": "./index.js"
 	}
